nothing of note

+++ src/test.cpp
@@ -33,6 +33,6 @@ int main(void){
     int8_t c = -120;
 
     //std::cout << (int)min(c) << N << valid_subtract<int8_t>(c, 1);
-    std::cout << testing(c);
+    std::cout << (int)testing(c);
 
 }
